Multithreading is here!
So, the big change is that now Universe splits object list into equal parts and processes them simultaneously. Huge performance boost! (Check the video)

Changed most variable types from double to float. And about that..

I added constants for Sun & Earth and did some testing with them, they dont seem to be simulated properly yet... I suspect extremely big or small numbers cause some odd stuff in the physics. Working on it...

I also reworked vertices a bit, so they now have sizes & color (doesn't work yet D:). Though, they may seem square, in fact, objects are round. Planning to code in tesselation later..

+++ Physics.h
@@ -2,7 +2,7 @@
 #include <vector>
 #include "ObjectHandler.h"
 
-float Force(PhysicalObject& Obj1, PhysicalObject& Obj2, const float& DistSq);
-float DistSq(PhysicalObject& Obj1, PhysicalObject& Obj2);
+double Force(PhysicalObject& Obj1, PhysicalObject& Obj2, const double& DistSq);
+double DistSq(PhysicalObject& Obj1, PhysicalObject& Obj2);
 void MoveObject(PhysicalObject& TargetObject, std::vector<PhysicalObject*>& ObjectReg, const int& timescale);
 void UpdatePhysics(std::vector<PhysicalObject*>& ObjectArray, const int& timescale);
